add header to auth layout and move navigation to components
