+++ web-app/messages/en.json
@@ -3,6 +3,12 @@
     "AgentCard": {
       "button": "Show details",
       "pricing": "{price, plural, =0 {For Free} =1 {1 credit} other {# credits}}"
+    },
+    "MainNavigation": {
+      "agents": "Agents Gallery",
+      "howItWorks": "How It Works",
+      "community": "Join Our Community",
+      "monetize": "Monetize"
     }
   },
   "Auth": {
@@ -104,14 +110,6 @@
         "template": "Sokosumi - %s"
       },
       "description": "Hire yourself an agent to finish the most time consuming tasks"
-    },
-    "Header": {
-      "Navigation": {
-        "agents": "Agents Gallery",
-        "howItWorks": "How It Works",
-        "community": "Join Our Community",
-        "monetize": "Monetize"
-      }
     },
     "Footer": {
       "Navigation": {
