+++ web-app/src/app/(auth)/layout.tsx
@@ -1,9 +1,8 @@
-import { X } from "lucide-react";
 import type { Metadata } from "next";
-import Link from "next/link";
 import { getTranslations } from "next-intl/server";
 
-import { Button } from "@/components/ui/button";
+import CloseButton from "./components/close-button";
+import Header from "./components/header";
 
 export async function generateMetadata(): Promise<Metadata> {
   const t = await getTranslations("Auth.Metadata");
@@ -23,14 +22,11 @@ export default function AuthLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <div className="flex flex-1 flex-col items-center justify-center">
+    <div className="flex flex-1 flex-col">
+      <Header />
       <div className="flex w-full max-w-2xl flex-1 flex-col gap-6 p-6">
         <div className="flex items-center justify-start">
-          <Link href="/">
-            <Button className="h-10 w-10 rounded-md bg-black">
-              <X className="text-md text-white" />
-            </Button>
-          </Link>
+          <CloseButton />
         </div>
         <div className="flex flex-1 items-center justify-center rounded-lg border border-gray-200 p-3">
           <div className="flex w-full flex-col">{children}</div>
